Move public methods to util
